Restructured App.tsx, added CameraPage and PermissionsPage
